Merge master to release branch in preparation for releasing 4.2.4

+++ DeepSkyStacker/Multitask.h
@@ -37,7 +37,7 @@ public :
 	{
 		HANDLE			hResult = nullptr;
 
-		for (LONG i = 0;(i<m_vThreadIds.size()) && !hResult;i++)
+		for (size_t i = 0;(i<m_vThreadIds.size()) && !hResult;i++)
 		{
 			if (dwThreadId == m_vThreadIds[i])
 				hResult = m_vEvents[i];
